Update MainFragment.kt : 파일 구조 수정
